+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamo/impl/kvstore/KVStorePlugin.java
@@ -41,7 +41,7 @@ public interface KVStorePlugin extends Plugin {
 	 * @param collection the collection
 	 * @return count of elements into collection
 	 */
-	long count(String collection);
+	int count(String collection);
 
 	/**
 	 * Adds an element defined by an id in a collection.
