+++ vertigo-dynamo-impl/src/test/java/io/vertigo/dynamo/kvstore/AbstractKVStoreManagerTest.java
@@ -54,7 +54,7 @@ public abstract class AbstractKVStoreManagerTest extends AbstractTestCaseJU4 {
 	@Test
 	public void testCount() {
 		try (VTransactionWritable transaction = transactionManager.createCurrentTransaction()) {
-			final long nbFlowers = kvStoreManager.count("flowers");
+			final int nbFlowers = kvStoreManager.count("flowers");
 			Assert.assertEquals(0, nbFlowers);
 			final Flower tulip1 = buildFlower("tulip", 100);
 			kvStoreManager.put("flowers", "1", tulip1);
@@ -66,17 +66,17 @@ public abstract class AbstractKVStoreManagerTest extends AbstractTestCaseJU4 {
 			kvStoreManager.put("flowers", "3", tulip3);
 
 			//count after 3 inserts
-			final long nbFlowers2 = kvStoreManager.count("flowers");
+			final int nbFlowers2 = kvStoreManager.count("flowers");
 			Assert.assertEquals(3, nbFlowers2);
 
 			//count after 1 more insert of same value
 			kvStoreManager.put("flowers", "4", tulip3);
-			final long nbFlowers3 = kvStoreManager.count("flowers");
+			final int nbFlowers3 = kvStoreManager.count("flowers");
 			Assert.assertEquals(4, nbFlowers3);
 
 			//count after 1 insert of same key : update
 			kvStoreManager.put("flowers", "3", tulip3);
-			final long nbFlowers4 = kvStoreManager.count("flowers");
+			final int nbFlowers4 = kvStoreManager.count("flowers");
 			Assert.assertEquals(4, nbFlowers4);
 		}
 	}
